feat: 옵션 선택 누락 수정 및 api 수정
