+++ src/pages/SocialLogin/NaverSocial.tsx
@@ -9,18 +9,6 @@ import { LOGIN_USER } from '../../slice/userSlice';
 const NaverSocial = () => {
 	const navigate = useNavigate();
 	const dispatch = useDispatch();
-	useEffect(() => {
-		if (document.cookie.includes('access')) {
-			const cookie = document.cookie;
-			const accessTokenMatch = cookie.match(/access_token=([^;]*)/);
-			const accessToken = accessTokenMatch ? accessTokenMatch[1] : null;
-
-			const refreshTokenMatch = cookie.match(/refresh_token=([^;]*)/);
-			const refreshToken = refreshTokenMatch ? refreshTokenMatch[1] : null;
-			if (accessToken) localStorage.setItem('access_token', accessToken);
-			if (refreshToken) localStorage.setItem('refresh_token', refreshToken);
-		}
-	}, []);
 
 	const loginCancelHandler = () => {
 		localToken.remove();
